Extract job list to separate component
